Refactor, split functions into separate modules and reorganize file structure
