+++ index.html
@@ -41,6 +41,6 @@
       </section>
     </main>
 
-    <script src="./js/index.js" type="module"></script>
+    <script src="./js/app.js" type="module"></script>
   </body>
 </html>
